Time: 475 ms (72.13%) | Memory: 243.6 MB (16.03%) - LeetSync

+++ 493-reverse-pairs/reverse-pairs.cpp
@@ -32,7 +32,7 @@ public:
         int right = mid + 1;
         int cnt = 0;
         for (int i=low; i<= mid;i++){
-            while(right <= high && (long long)arr[i] > 2LL*arr[right]){
+            while(right <= high && (long long)arr[i] > 2*(long long)arr[right]){
                 right++;
             }
             cnt += (right-(mid+1));
